Added 1707. Maximum XOR With an Element From Array
The problem "Maximum XOR With an Element From Array" involves finding the maximum XOR value obtainable by XORing an integer from a given array with a specified element. This is efficiently tackled using a Trie data structure, which allows for easy manipulation of bits. By inserting the binary representations of the integers into the Trie, we can quickly traverse the structure to find the optimal combinations that maximize the XOR value with minimal computational overhead.